FLAVIA prealpha v0.2
added worker classes for performance boosts and various features, such as tracing, monitoring, config, logging....
die worker .py codes should be in a folder labeled "backend", while the gui features should be in a folder labeled "gui" in the same directory. the main.py starts the entire thing. 
the py caches can be used or ignored.